+++ saleor/graphql/account/mutations/base.py
@@ -25,6 +25,7 @@ from ...account.types import Address, AddressInput, User
 from ...app.dataloaders import get_app_promise
 from ...channel.utils import clean_channel, validate_channel
 from ...core.context import set_mutation_flag_in_context
+from ...core.descriptions import ADDED_IN_310
 from ...core.enums import LanguageCodeEnum
 from ...core.mutations import (
     BaseMutation,
@@ -420,6 +421,9 @@ class UserAddressInput(graphene.InputObjectType):
 class CustomerInput(UserInput, UserAddressInput):
     language_code = graphene.Field(
         LanguageCodeEnum, required=False, description="User language code."
+    )
+    external_reference = graphene.String(
+        description="External ID of the customer." + ADDED_IN_310, required=False
     )