+++ saleor/graphql/account/tests/test_account.py
@@ -622,10 +622,11 @@ def test_query_staff_user_with_orders_and_manage_orders_perm(
 
 
 USER_QUERY = """
-    query User($id: ID $email: String) {
-        user(id: $id, email: $email) {
+    query User($id: ID $email: String, $externalReference: String) {
+        user(id: $id, email: $email, externalReference: $externalReference) {
             id
             email
+            externalReference
         }
     }
 """
@@ -642,6 +643,27 @@ def test_query_user_by_email_address(
     content = get_graphql_content(response)
     data = content["data"]["user"]
     assert customer_user.email == data["email"]
+
+
+def test_query_user_by_external_reference(
+    user_api_client, customer_user, permission_manage_users
+):
+    # given
+    user = customer_user
+    ext_ref = "test-ext-ref"
+    user.external_reference = ext_ref
+    user.save(update_fields=["external_reference"])
+    variables = {"externalReference": ext_ref}
+
+    # when
+    response = user_api_client.post_graphql(
+        USER_QUERY, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["user"]
+    assert data["externalReference"] == user.external_reference
 
 
 def test_query_user_by_id_and_email(
@@ -1510,7 +1532,9 @@ CUSTOMER_CREATE_MUTATION = """
     mutation CreateCustomer(
         $email: String, $firstName: String, $lastName: String, $channel: String
         $note: String, $billing: AddressInput, $shipping: AddressInput,
-        $redirect_url: String, $languageCode: LanguageCodeEnum) {
+        $redirect_url: String, $languageCode: LanguageCodeEnum,
+        $externalReference: String
+    ) {
         customerCreate(input: {
             email: $email,
             firstName: $firstName,
@@ -1521,6 +1545,7 @@ CUSTOMER_CREATE_MUTATION = """
             redirectUrl: $redirect_url,
             languageCode: $languageCode,
             channel: $channel,
+            externalReference: $externalReference
         }) {
             errors {
                 field
@@ -1542,6 +1567,7 @@ CUSTOMER_CREATE_MUTATION = """
                 isActive
                 isStaff
                 note
+                externalReference
             }
         }
     }
@@ -1566,6 +1592,7 @@ def test_customer_create(
     note = "Test user"
     address_data = convert_dict_keys_to_camel_case(address.as_data())
     redirect_url = "https://www.example.com"
+    external_reference = "test-ext-ref"
     variables = {
         "email": email,
         "firstName": first_name,
@@ -1576,6 +1603,7 @@ def test_customer_create(
         "redirect_url": redirect_url,
         "languageCode": "PL",
         "channel": channel_PLN.slug,
+        "externalReference": external_reference,
     }
 
     response = staff_api_client.post_graphql(
@@ -1600,6 +1628,7 @@ def test_customer_create(
     assert data["user"]["lastName"] == last_name
     assert data["user"]["note"] == note
     assert data["user"]["languageCode"] == "PL"
+    assert data["user"]["externalReference"] == external_reference
     assert not data["user"]["isStaff"]
     assert data["user"]["isActive"]
 
@@ -1744,6 +1773,29 @@ def test_customer_create_with_upper_case_email(
     assert data["user"]["email"] == email.lower()
 
 
+def test_customer_create_with_non_unique_external_reference(
+    staff_api_client, permission_manage_users, customer_user
+):
+    # given
+    ext_ref = "test-ext-ref"
+    customer_user.external_reference = ext_ref
+    customer_user.save(update_fields=["external_reference"])
+
+    variables = {"email": "mail.test@exampale.com", "externalReference": ext_ref}
+
+    # when
+    response = staff_api_client.post_graphql(
+        CUSTOMER_CREATE_MUTATION, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    error = content["data"]["customerCreate"]["errors"][0]
+    assert error["field"] == "externalReference"
+    assert error["code"] == AccountErrorCode.UNIQUE.name
+    assert error["message"] == "User with this External reference already exists."
+
+
 def test_customer_update(
     staff_api_client, staff_user, customer_user, address, permission_manage_users
 ):
@@ -1751,16 +1803,22 @@ def test_customer_update(
     mutation UpdateCustomer(
             $id: ID!, $firstName: String, $lastName: String,
             $isActive: Boolean, $note: String, $billing: AddressInput,
-            $shipping: AddressInput, $languageCode: LanguageCodeEnum) {
-        customerUpdate(id: $id, input: {
-            isActive: $isActive,
-            firstName: $firstName,
-            lastName: $lastName,
-            note: $note,
-            defaultBillingAddress: $billing
-            defaultShippingAddress: $shipping,
-            languageCode: $languageCode
-        }) {
+            $shipping: AddressInput, $languageCode: LanguageCodeEnum,
+            $externalReference: String
+        ) {
+        customerUpdate(
+            id: $id,
+            input: {
+                isActive: $isActive,
+                firstName: $firstName,
+                lastName: $lastName,
+                note: $note,
+                defaultBillingAddress: $billing
+                defaultShippingAddress: $shipping,
+                languageCode: $languageCode,
+                externalReference: $externalReference
+                }
+            ) {
             errors {
                 field
                 message
@@ -1778,6 +1836,7 @@ def test_customer_update(
                 languageCode
                 isActive
                 note
+                externalReference
             }
         }
     }
@@ -1794,6 +1853,7 @@ def test_customer_update(
     first_name = "new_first_name"
     last_name = "new_last_name"
     note = "Test update note"
+    external_reference = "test-ext-ref"
     address_data = convert_dict_keys_to_camel_case(address.as_data())
 
     new_street_address = "Updated street address"
@@ -1808,6 +1868,7 @@ def test_customer_update(
         "billing": address_data,
         "shipping": address_data,
         "languageCode": "PL",
+        "externalReference": external_reference,
     }
     response = staff_api_client.post_graphql(
         query, variables, permissions=[permission_manage_users]
@@ -1833,6 +1894,7 @@ def test_customer_update(
     assert data["user"]["lastName"] == last_name
     assert data["user"]["note"] == note
     assert data["user"]["languageCode"] == "PL"
+    assert data["user"]["externalReference"] == external_reference
     assert not data["user"]["isActive"]
 
     (
@@ -1857,6 +1919,130 @@ def test_customer_update(
         generate_address_search_document_value(shipping_address)
         in customer_user.search_document
     )
+
+
+UPDATE_CUSTOMER_BY_EXTERNAL_REFERENCE = """
+    mutation UpdateCustomer(
+        $id: ID, $externalReference: String, $input: CustomerInput!
+    ) {
+        customerUpdate(id: $id, externalReference: $externalReference, input: $input) {
+            errors {
+                field
+                message
+                code
+            }
+            user {
+                id
+                externalReference
+                firstName
+            }
+        }
+    }
+    """
+
+
+def test_customer_update_by_external_reference(
+    staff_api_client, customer_user, permission_manage_users
+):
+    # given
+    query = UPDATE_CUSTOMER_BY_EXTERNAL_REFERENCE
+    user = customer_user
+    new_name = "updated name"
+    ext_ref = "test-ext-ref"
+    user.external_reference = ext_ref
+    user.save(update_fields=["external_reference"])
+
+    variables = {
+        "externalReference": ext_ref,
+        "input": {"firstName": new_name},
+    }
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    user.refresh_from_db()
+    data = content["data"]["customerUpdate"]
+    assert not data["errors"]
+    assert data["user"]["firstName"] == new_name == user.first_name
+    assert data["user"]["id"] == graphene.Node.to_global_id("User", user.id)
+    assert data["user"]["externalReference"] == ext_ref
+
+
+def test_update_customer_by_both_id_and_external_reference(
+    staff_api_client, customer_user, permission_manage_users
+):
+    # given
+    query = UPDATE_CUSTOMER_BY_EXTERNAL_REFERENCE
+    variables = {"input": {}, "externalReference": "whatever", "id": "whatever"}
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["customerUpdate"]
+    assert not data["user"]
+    assert (
+        data["errors"][0]["message"]
+        == "Argument 'id' cannot be combined with 'external_reference'"
+    )
+
+
+def test_update_customer_by_external_reference_not_existing(
+    staff_api_client, customer_user, permission_manage_users
+):
+    # given
+    query = UPDATE_CUSTOMER_BY_EXTERNAL_REFERENCE
+    ext_ref = "non-existing-ext-ref"
+    variables = {
+        "input": {},
+        "externalReference": ext_ref,
+    }
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["customerUpdate"]
+    assert not data["user"]
+    assert data["errors"][0]["message"] == f"Couldn't resolve to a node: {ext_ref}"
+    assert data["errors"][0]["field"] == "externalReference"
+
+
+def test_update_customer_with_non_unique_external_reference(
+    staff_api_client, permission_manage_users, user_list
+):
+    # given
+    query = UPDATE_CUSTOMER_BY_EXTERNAL_REFERENCE
+
+    ext_ref = "test-ext-ref"
+    user_1 = user_list[0]
+    user_1.external_reference = ext_ref
+    user_1.save(update_fields=["external_reference"])
+    user_2_id = graphene.Node.to_global_id("User", user_list[1].id)
+
+    variables = {"input": {"externalReference": ext_ref}, "id": user_2_id}
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    error = content["data"]["customerUpdate"]["errors"][0]
+    assert error["field"] == "externalReference"
+    assert error["code"] == AccountErrorCode.UNIQUE.name
+    assert error["message"] == "User with this External reference already exists."
 
 
 UPDATE_CUSTOMER_EMAIL_MUTATION = """
@@ -2560,14 +2746,15 @@ def test_account_delete_other_customer_token(user_api_client):
 
 
 CUSTOMER_DELETE_MUTATION = """
-    mutation CustomerDelete($id: ID!) {
-        customerDelete(id: $id){
+    mutation CustomerDelete($id: ID, $externalReference: String) {
+        customerDelete(id: $id, externalReference: $externalReference) {
             errors {
                 field
                 message
             }
             user {
                 id
+                externalReference
             }
         }
     }
@@ -2696,6 +2883,72 @@ def test_customer_delete_errors(customer_user, admin_user, staff_user):
 
     # should not raise any errors
     CustomerDelete.clean_instance(info, customer_user)
+
+
+def test_customer_delete_by_external_reference(
+    staff_api_client, customer_user, permission_manage_users
+):
+    # given
+    user = customer_user
+    query = CUSTOMER_DELETE_MUTATION
+    ext_ref = "test-ext-ref"
+    user.external_reference = ext_ref
+    user.save(update_fields=["external_reference"])
+    variables = {"externalReference": ext_ref}
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["customerDelete"]
+    with pytest.raises(user._meta.model.DoesNotExist):
+        user.refresh_from_db()
+    assert not data["errors"]
+    assert data["user"]["externalReference"] == ext_ref
+    assert data["user"]["id"] == graphene.Node.to_global_id("User", user.id)
+
+
+def test_delete_customer_by_both_id_and_external_reference(
+    staff_api_client, customer_user, permission_manage_users
+):
+    # given
+    query = CUSTOMER_DELETE_MUTATION
+    variables = {"externalReference": "whatever", "id": "whatever"}
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    errors = content["data"]["customerDelete"]["errors"]
+    assert (
+        errors[0]["message"]
+        == "Argument 'id' cannot be combined with 'external_reference'"
+    )
+
+
+def test_delete_customer_by_external_reference_not_existing(
+    staff_api_client, customer_user, permission_manage_users
+):
+    # given
+    query = CUSTOMER_DELETE_MUTATION
+    ext_ref = "non-existing-ext-ref"
+    variables = {"externalReference": ext_ref}
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_users]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    errors = content["data"]["customerDelete"]["errors"]
+    assert errors[0]["message"] == f"Couldn't resolve to a node: {ext_ref}"
 
 
 STAFF_CREATE_MUTATION = """