+++ saleor/graphql/attribute/mutations/attribute_delete.py
@@ -2,15 +2,20 @@ import graphene
 
 from ....attribute import models as models
 from ....core.permissions import ProductTypePermissions
-from ...core.mutations import ModelDeleteMutation
+from ...core.descriptions import ADDED_IN_310
+from ...core.mutations import ModelDeleteMutation, ModelWithExtRefMutation
 from ...core.types import AttributeError
 from ...plugins.dataloaders import get_plugin_manager_promise
 from ..types import Attribute
 
 
-class AttributeDelete(ModelDeleteMutation):
+class AttributeDelete(ModelDeleteMutation, ModelWithExtRefMutation):
     class Arguments:
-        id = graphene.ID(required=True, description="ID of an attribute to delete.")
+        id = graphene.ID(required=False, description="ID of an attribute to delete.")
+        external_reference = graphene.String(
+            required=False,
+            description=f"External ID of an attribute to delete. {ADDED_IN_310}",
+        )
 
     class Meta:
         model = models.Attribute