+++ saleor/graphql/order/tests/mutations/test_draft_order_update.py
@@ -86,14 +86,17 @@ def test_draft_order_update_voucher_not_available(
 
 DRAFT_ORDER_UPDATE_MUTATION = """
     mutation draftUpdate(
-        $id: ID!, $voucher: ID!, $customerNote: String, $shippingAddress: AddressInput
+        $id: ID!, $voucher: ID!, $customerNote: String, $shippingAddress: AddressInput,
+        $externalReference: String
     ) {
-        draftOrderUpdate(id: $id,
-                            input: {
-                                voucher: $voucher,
-                                customerNote: $customerNote,
-                                shippingAddress: $shippingAddress,
-                            }) {
+        draftOrderUpdate(
+            id: $id,
+            input: {
+                voucher: $voucher,
+                customerNote: $customerNote,
+                shippingAddress: $shippingAddress,
+                externalReference: $externalReference
+            }) {
             errors {
                 field
                 message
@@ -101,6 +104,7 @@ DRAFT_ORDER_UPDATE_MUTATION = """
             }
             order {
                 userEmail
+                externalReference
             }
         }
     }
@@ -117,10 +121,12 @@ def test_draft_order_update(
     order_id = graphene.Node.to_global_id("Order", order.id)
     voucher_id = graphene.Node.to_global_id("Voucher", voucher.id)
     customer_note = "Test customer note"
+    external_reference = "test-ext-ref"
     variables = {
         "id": order_id,
         "voucher": voucher_id,
         "customerNote": customer_note,
+        "externalReference": external_reference,
     }
 
     response = staff_api_client.post_graphql(
@@ -133,6 +139,11 @@ def test_draft_order_update(
     assert order.voucher
     assert order.customer_note == customer_note
     assert order.search_vector
+    assert (
+        data["order"]["externalReference"]
+        == external_reference
+        == order.external_reference
+    )
 
 
 def test_draft_order_update_with_non_draft_order(
@@ -335,3 +346,146 @@ def test_draft_order_update_assign_user_when_existing_customer_email_provided(
     assert not data["errors"]
     assert order.user == user
     assert order.user_email == user_email
+
+
+DRAFT_ORDER_UPDATE_BY_EXTERNAL_REFERENCE = """
+    mutation draftUpdate(
+        $id: ID
+        $externalReference: String
+        $input: DraftOrderInput!
+    ) {
+        draftOrderUpdate(
+            id: $id
+            externalReference: $externalReference
+            input: $input
+        ) {
+            errors {
+                field
+                message
+                code
+            }
+            order {
+                id
+                externalReference
+                voucher {
+                    id
+                }
+            }
+        }
+    }
+    """
+
+
+def test_draft_order_update_by_external_reference(
+    staff_api_client, permission_manage_orders, draft_order, voucher_free_shipping
+):
+    # given
+    query = DRAFT_ORDER_UPDATE_BY_EXTERNAL_REFERENCE
+
+    order = draft_order
+    assert not order.voucher
+    voucher = voucher_free_shipping
+    voucher_id = graphene.Node.to_global_id("Voucher", voucher.id)
+    ext_ref = "test-ext-ref"
+    order.external_reference = ext_ref
+    order.save(update_fields=["external_reference"])
+
+    variables = {
+        "externalReference": ext_ref,
+        "input": {"voucher": voucher_id},
+    }
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["draftOrderUpdate"]
+    assert not data["errors"]
+    assert data["order"]["externalReference"] == ext_ref
+    assert data["order"]["id"] == graphene.Node.to_global_id("Order", order.id)
+    assert data["order"]["voucher"]["id"] == voucher_id
+    order.refresh_from_db()
+    assert order.voucher
+
+
+def test_draft_order_update_by_both_id_and_external_reference(
+    staff_api_client, permission_manage_orders, voucher_free_shipping
+):
+    # given
+    query = DRAFT_ORDER_UPDATE_BY_EXTERNAL_REFERENCE
+
+    variables = {
+        "id": "test-id",
+        "externalReference": "test-ext-ref",
+        "input": {},
+    }
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["draftOrderUpdate"]
+    assert not data["order"]
+    assert (
+        data["errors"][0]["message"]
+        == "Argument 'id' cannot be combined with 'external_reference'"
+    )
+
+
+def test_draft_order_update_by_external_reference_not_existing(
+    staff_api_client, permission_manage_orders, voucher_free_shipping
+):
+    # given
+    query = DRAFT_ORDER_UPDATE_BY_EXTERNAL_REFERENCE
+    ext_ref = "non-existing-ext-ref"
+    variables = {
+        "externalReference": ext_ref,
+        "input": {},
+    }
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["draftOrderUpdate"]
+    assert not data["order"]
+    assert data["errors"][0]["message"] == f"Couldn't resolve to a node: {ext_ref}"
+
+
+def test_draft_order_update_with_non_unique_external_reference(
+    staff_api_client,
+    permission_manage_orders,
+    draft_order,
+    order_list,
+):
+    # given
+    query = DRAFT_ORDER_UPDATE_BY_EXTERNAL_REFERENCE
+
+    draft_order_id = graphene.Node.to_global_id("Order", draft_order.pk)
+    ext_ref = "test-ext-ref"
+    order = order_list[1]
+    order.external_reference = ext_ref
+    order.save(update_fields=["external_reference"])
+
+    variables = {"id": draft_order_id, "input": {"externalReference": ext_ref}}
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+
+    # then
+    error = content["data"]["draftOrderUpdate"]["errors"][0]
+    assert error["field"] == "externalReference"
+    assert error["code"] == OrderErrorCode.UNIQUE.name
+    assert error["message"] == "Order with this External reference already exists."